Fix error when updating image in admin and price format

+++ controllers/product.js
@@ -82,14 +82,18 @@ exports.read = async (req, res) => {
 exports.update = async (req, res) => {
 	const productId = req.params.productId;
 
-	req.body.fileName = req.file.filename;
+	if (req.file !== undefined) {
+		req.body.fileName = req.file.filename;
+	}
 
 	const oldProduct = await Product.findByIdAndUpdate(productId, req.body);
 
-	fs.unlink(`uploads/${oldProduct.fileName}`, err => {
-		if (err) throw err;
-		console.log('Image successfully deleted from the filesystem');
-	});
+	if (req.file !== undefined && req.file.filename !== oldProduct.fileName) {
+		fs.unlink(`uploads/${oldProduct.fileName}`, err => {
+			if (err) throw err;
+			console.log('Image deleted from the filesystem');
+		});
+	}
 
 	res.json({
 		successMessage: 'Product successfully updated',
